pyproject.toml
[project]
name = "ohheybrian"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "alembic==1.11.1 ; python_version >= '3.8' and python_version < '4.0'",
    "blinker==1.6.2 ; python_version >= '3.8' and python_version < '4.0'",
    "click==8.1.3 ; python_version >= '3.8' and python_version < '4.0'",
    "colorama==0.4.6 ; python_version >= '3.8' and python_version < '4.0' and platform_system == 'Windows'",
    "flask==2.3.2 ; python_version >= '3.8' and python_version < '4.0'",
    "flask-migrate==4.0.4 ; python_version >= '3.8' and python_version < '4.0'",
    "flask-sqlalchemy>=3.1.1",
    "greenlet==2.0.2 ; (python_version >= '3.8' and python_version < '4.0' and platform_machine == 'aarch64') or (python_version >= '3.8' and python_version < '4.0' and platform_machine == 'ppc64le') or (python_version >= '3.8' and python_version < '4.0' and platform_machine == 'x86_64') or (python_version >= '3.8' and python_version < '4.0' and platform_machine == 'amd64') or (python_version >= '3.8' and python_version < '4.0' and platform_machine == 'AMD64') or (python_version >= '3.8' and python_version < '4.0' and platform_machine == 'win32') or (python_version >= '3.8' and python_version < '4.0' and platform_machine == 'WIN32')",
    "htmx-flask==0.1.0 ; python_version >= '3.8' and python_version < '4.0'",
    "importlib-metadata==6.6.0 ; python_version >= '3.8' and python_version < '3.10'",
    "importlib-resources==5.12.0 ; python_version >= '3.8' and python_version < '3.9'",
    "itsdangerous==2.1.2 ; python_version >= '3.8' and python_version < '4.0'",
    "jinja-partials==0.1.1 ; python_version >= '3.8' and python_version < '4.0'",
    "jinja2==3.1.2 ; python_version >= '3.8' and python_version < '4.0'",
    "mako==1.2.4 ; python_version >= '3.8' and python_version < '4.0'",
    "markupsafe==2.1.2 ; python_version >= '3.8' and python_version < '4.0'",
    "marshmallow==3.19.0 ; python_version >= '3.8' and python_version < '4.0'",
    "nh3==0.2.13 ; python_version >= '3.8' and python_version < '4.0'",
    "packaging==23.1 ; python_version >= '3.8' and python_version < '4.0'",
    "python-dotenv==1.0.0 ; python_version >= '3.8' and python_version < '4.0'",
    "sqlalchemy==2.0.16 ; python_version >= '3.8' and python_version < '4.0'",
    "typing-extensions==4.6.3 ; python_version >= '3.8' and python_version < '4.0'",
    "webargs==8.4.0",
    "werkzeug==2.3.4 ; python_version >= '3.8' and python_version < '4.0'",
    "zipp==3.15.0 ; python_version >= '3.8' and python_version < '3.10'",
    "flask-login>=0.6.3",
    "flask-cors>=4.0.0",
    "python-slugify>=8.0.4",
    "markdown>=3.10",
    "feedgen>=1.0.0",
    "markdownify>=1.2.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "flask-shell-ipython>=0.5.3",
]

[tool.hatch.metadata]
allow-direct-references = true
